Move workshop to a new package
