+++ 06-workshop/src/main/scala/dagmendez/workshop/b/ApplyAndFrom.scala
