+++ 06-workshop/src/main/scala/dagmendez/workshop/c/CustomErrors.scala
