+++ 06-workshop/src/main/scala/dagmendez/workshop/d/FromApi.scala
